Armageddon
Too many features and fixes to keep track of

+++ BPW2/Assets/01_Scripts/00_Player/PlayerStatus.cs
@@ -53,7 +53,7 @@ public class PlayerStatus : MonoBehaviour
 
     public void ChangeMaxHealth(int newMax)
     {
-        healthBar.UpdateBarMaxValue(newMax/maxHealth);
+        healthBar.UpdateBarMaxValue((float)newMax / (float)maxHealth);
         maxHealth = newMax;
         health = Mathf.Clamp(health, 0, maxHealth);
         UpdateHealthBar();
